atomized model, import export atomized model

+++ test/chain/test_chain_import_export.py
@@ -9,12 +9,17 @@ from core.composer.chain import Chain
 from core.composer.node import PrimaryNode, SecondaryNode
 from core.models.data import InputData
 from utilities.synthetic.chain_template_new import ChainTemplate, extract_subtree_root
+from test.chain.test_atomized_chain import create_chain_with_several_chain_models_nested
 
 
 @pytest.fixture(scope="session", autouse=True)
 def creation_model_files_before_after_tests(request):
+    # Check for exists files if last time tests crash
+    delete_json_models_files()
+
     create_json_models_files()
     request.addfinalizer(delete_json_models_files)
+    request.addfinalizer(delete_folder_with_nested_model_json)
 
 
 def create_json_models_files():
@@ -31,18 +36,35 @@ def create_json_models_files():
     chain_empty.save_chain("data/test_empty_chain_convert_to_json.json")
 
 
+def delete_folder_with_nested_model_json():
+    """
+    Delete folders.
+    """
+    folders_name = ["data/chain_model_1",
+                    "data/check_load_model_chain_1"]
+
+    for folder_path in folders_name:
+        dir_path = os.path.abspath(folder_path)
+        shutil.rmtree(dir_path)
+
+
 def delete_json_models_files():
     """
     Delete JSON's files.
     """
-    with open("data/test_fitted_chain_convert_to_json.json", 'r') as json_file:
-        chain_fitted_object = json.load(json_file)
+    if os.path.isfile("data/test_fitted_chain_convert_to_json.json"):
+        with open("data/test_fitted_chain_convert_to_json.json", 'r') as json_file:
+            chain_fitted_object = json.load(json_file)
 
-    delete_fitted_models(chain_fitted_object)
+        delete_fitted_models(chain_fitted_object)
 
-    os.remove("data/test_fitted_chain_convert_to_json.json")
-    os.remove("data/test_empty_chain_convert_to_json.json")
-    os.remove("data/test_chain_convert_to_json.json")
+    files_name = ["data/test_fitted_chain_convert_to_json.json",
+                  "data/test_empty_chain_convert_to_json.json",
+                  "data/test_chain_convert_to_json.json"]
+
+    for file in files_name:
+        if os.path.isfile(file):
+            os.remove(file)
 
 
 def delete_fitted_models(chain):
@@ -51,9 +73,11 @@ def delete_fitted_models(chain):
 
     :param chain: chain which model's need to delete
     """
-    model_path = chain['nodes'][0]['trained_model_path']
-    dir_path = os.path.dirname(os.path.abspath(model_path))
-    shutil.rmtree(dir_path)
+    root_node = chain['nodes'][0]
+    if 'trained_model_path' in root_node:
+        model_path = root_node['trained_model_path']
+        dir_path = os.path.dirname(os.path.abspath(model_path))
+        shutil.rmtree(dir_path)
 
 
 def create_chain() -> Chain:
@@ -229,7 +253,7 @@ def test_import_custom_json_object_to_chain_and_fit_correctly_no_exception():
     train_data = InputData.from_csv(train_file_path)
 
     data_path = str(os.path.dirname(__file__))
-    json_file_path = os.path.join(data_path, '..', 'test', 'data', 'test_custom_json_template.json')
+    json_file_path = os.path.join(data_path, '..', 'data', 'test_custom_json_template.json')
 
     chain = Chain()
     chain_template = ChainTemplate(chain)
@@ -279,3 +303,20 @@ def test_extract_subtree_root():
     assertion_list = [True if expected_types[index] == actual_types[index] else False
                       for index in range(len(expected_types))]
     assert all(assertion_list)
+
+
+def test_atomized_chain_import_export_correctly():
+    chain = create_chain_with_several_chain_models_nested()
+    chain.save_chain("data/chain_model_1.json")
+
+    train_file_path, test_file_path = get_scoring_case_data_paths()
+    train_data = InputData.from_csv(train_file_path)
+
+    chain = Chain()
+    chain.load_chain("data/chain_model_1.json")
+    chain.fit(train_data)
+    json_actual = chain.save_chain("data/check_load_model_chain_1.json")
+
+    delete_fitted_models(json.loads(json_actual))
+    os.remove("data/chain_model_1.json")
+    os.remove("data/check_load_model_chain_1.json")
